+++ api/server.go
@@ -1,3 +1,15 @@
+// Copyright (c) 2022 Gobalsky Labs Limited
+//
+// Use of this software is governed by the Business Source License included
+// in the LICENSE file and at https://www.mariadb.com/bsl11.
+//
+// Change Date: 18 months from the later of the date of the first publicly
+// available Distribution of this version of the repository, and 25 June 2022.
+//
+// On the date above, in accordance with the Business Source License, use
+// of this software will be governed by version 3 or later of the GNU General
+// Public License.
+
 package api
 
 import (
@@ -463,6 +475,10 @@ func (g *GRPCServer) Start(ctx context.Context, lis net.Listener) error {
 		positionService:      g.positionServiceV2,
 		accountService:       g.accountServiceV2,
 		rewardService:        g.rewardServiceV2,
+		depositService:       g.depositServiceV2,
+		withdrawalService:    g.withdrawalServiceV2,
+		oracleSpecService:    g.oracleSpecServiceV2,
+		oracleDataService:    g.oracleDataServiceV2,
 	}
 
 	protoapi2.RegisterTradingDataServiceServer(g.srv, tradingDataSvcV2)