+++ broker/broker.go
@@ -1,3 +1,15 @@
+// Copyright (c) 2022 Gobalsky Labs Limited
+//
+// Use of this software is governed by the Business Source License included
+// in the LICENSE file and at https://www.mariadb.com/bsl11.
+//
+// Change Date: 18 months from the later of the date of the first publicly
+// available Distribution of this version of the repository, and 25 June 2022.
+//
+// On the date above, in accordance with the Business Source License, use
+// of this software will be governed by version 3 or later of the GNU General
+// Public License.
+
 package broker
 
 import (
@@ -8,7 +20,7 @@ import (
 
 	"code.vegaprotocol.io/data-node/entities"
 	"code.vegaprotocol.io/data-node/logging"
-	types "code.vegaprotocol.io/protos/vega"
+	"code.vegaprotocol.io/protos/vega"
 	"code.vegaprotocol.io/vega/events"
 )
 
@@ -54,18 +66,17 @@ type ChainInfoI interface {
 	GetChainID() (string, error)
 }
 
-type OrderEvent interface {
-	events.Event
-	Order() *types.Order
-}
-
 type OrderEventWithVegaTime struct {
-	OrderEvent
+	events.Order
 	vegaTime time.Time
 }
 
-func (oe OrderEventWithVegaTime) VegaTime() time.Time {
+func (oe *OrderEventWithVegaTime) VegaTime() time.Time {
 	return oe.vegaTime
+}
+
+func (oe *OrderEventWithVegaTime) GetOrder() *vega.Order {
+	return oe.Order.Order()
 }
 
 // Broker - the base broker type
@@ -169,8 +180,8 @@ func (b *Broker) startSending(t events.Type, evt events.Event) {
 	}
 
 	if t == events.OrderEvent {
-		orderEvent := evt.(OrderEvent)
-		evt = OrderEventWithVegaTime{orderEvent, b.vegaTime}
+		orderEvent := evt.(*events.Order)
+		evt = &OrderEventWithVegaTime{*orderEvent, b.vegaTime}
 	}
 
 	ch <- []events.Event{evt}