+++ entities/asset.go
@@ -1,9 +1,22 @@
+// Copyright (c) 2022 Gobalsky Labs Limited
+//
+// Use of this software is governed by the Business Source License included
+// in the LICENSE file and at https://www.mariadb.com/bsl11.
+//
+// Change Date: 18 months from the later of the date of the first publicly
+// available Distribution of this version of the repository, and 25 June 2022.
+//
+// On the date above, in accordance with the Business Source License, use
+// of this software will be governed by version 3 or later of the GNU General
+// Public License.
+
 package entities
 
 import (
 	"fmt"
 	"time"
 
+	v2 "code.vegaprotocol.io/protos/data-node/api/v2"
 	pb "code.vegaprotocol.io/protos/vega"
 	"github.com/shopspring/decimal"
 )
@@ -56,4 +69,15 @@ func (a Asset) ToProto() *pb.Asset {
 	}
 
 	return pbAsset
+}
+
+func (a Asset) Cursor() *Cursor {
+	return NewCursor(a.ID.String())
+}
+
+func (a Asset) ToProtoEdge(_ ...any) *v2.AssetEdge {
+	return &v2.AssetEdge{
+		Node:   a.ToProto(),
+		Cursor: a.Cursor().Encode(),
+	}
 }