+++ entities/withdrawal.go
@@ -1,3 +1,15 @@
+// Copyright (c) 2022 Gobalsky Labs Limited
+//
+// Use of this software is governed by the Business Source License included
+// in the LICENSE file and at https://www.mariadb.com/bsl11.
+//
+// Change Date: 18 months from the later of the date of the first publicly
+// available Distribution of this version of the repository, and 25 June 2022.
+//
+// On the date above, in accordance with the Business Source License, use
+// of this software will be governed by version 3 or later of the GNU General
+// Public License.
+
 package entities
 
 import (
@@ -109,7 +121,8 @@ type WithdrawalCursor struct {
 func (wc WithdrawalCursor) String() string {
 	bs, err := json.Marshal(wc)
 	if err != nil {
-		return fmt.Sprintf(`{"vegaTime":%d,"id":"%s"}`, wc.VegaTime.Unix(), wc.ID)
+		// This should never happen
+		panic(fmt.Errorf("failed to marshal withdrawal cursor: %w", err))
 	}
 	return string(bs)
 }